update constant and state check (skip downloaded files)

+++ utils/process_assets.py
@@ -1,37 +1,53 @@
 import os
 from urllib.parse import urlparse
-from constants import LINK_ASSET_URL, FILE_ASSET_URL
+from constants import LINK_ASSET_URL, FILE_ASSET_URL, logger
 
-def download_supplementary_assets(udemy, assets, download_folder_path, course_id, lecture_id):
+
+def download_supplementary_assets(
+    udemy, assets, download_folder_path, course_id, lecture_id
+):
     for asset in assets:
-        match asset['asset_type']:
-            case 'File':
+        match asset["asset_type"]:
+            case "File":
                 process_files(udemy, asset, course_id, lecture_id, download_folder_path)
-            case 'ExternalLink':
-                process_external_links(udemy, asset, course_id, lecture_id, download_folder_path)
+            case "ExternalLink":
+                process_external_links(
+                    udemy, asset, course_id, lecture_id, download_folder_path
+                )
             case _:
                 pass
                 # Unsupported asset type. Please create a github issue if you'd like to add support for other types
 
-def process_files(udemy, asset, course_id, lecture_id, download_folder_path):
 
+def process_files(udemy, asset, course_id, lecture_id, download_folder_path):
     assets_folder = os.path.join(download_folder_path, "assets")
     if not os.path.exists(assets_folder):
         os.makedirs(assets_folder)
-    
-    asset_file_path = os.path.join(assets_folder, asset['filename'])
 
-    file_response = udemy.request(udemy.request(FILE_ASSET_URL.format(course_id=course_id, lecture_id=lecture_id, asset_id=asset['id'])).json()['download_urls']['File'][0]['file'])
+    asset_file_path = os.path.join(assets_folder, asset["filename"])
+
+    # Skip if file already exists and is non-empty
+    if os.path.isfile(asset_file_path) and os.path.getsize(asset_file_path) > 0:
+        logger.info(f"Skipped asset (already exists): {asset['filename']}")
+        return
+
+    file_response = udemy.request(
+        udemy.request(
+            FILE_ASSET_URL.format(
+                course_id=course_id, lecture_id=lecture_id, asset_id=asset["id"]
+            )
+        ).json()["download_urls"]["File"][0]["file"]
+    )
 
     file_response.raise_for_status()
 
-    with open(asset_file_path, 'wb') as file:
+    with open(asset_file_path, "wb") as file:
         for chunk in file_response.iter_content(chunk_size=8192):
             if chunk:
                 file.write(chunk)
 
-def process_external_links(udemy, asset, course_id, lecture_id, download_folder_path):
 
+def process_external_links(udemy, asset, course_id, lecture_id, download_folder_path):
     external_links_folder = os.path.join(download_folder_path, "external-links")
     if not os.path.exists(external_links_folder):
         os.makedirs(external_links_folder)
@@ -39,9 +55,18 @@ def process_external_links(udemy, asset, course_id, lecture_id, download_folder_
     asset_filename = f"{asset['filename']}.url"
     asset_file_path = os.path.join(external_links_folder, asset_filename)
 
-    response = udemy.request(LINK_ASSET_URL.format(course_id=course_id, lecture_id=lecture_id, asset_id=asset['id'])).json()
+    # Skip if link file already exists
+    if os.path.isfile(asset_file_path) and os.path.getsize(asset_file_path) > 0:
+        logger.info(f"Skipped external link (already exists): {asset_filename}")
+        return
 
-    asset_url = response['external_url']
+    response = udemy.request(
+        LINK_ASSET_URL.format(
+            course_id=course_id, lecture_id=lecture_id, asset_id=asset["id"]
+        )
+    ).json()
 
-    with open(asset_file_path, 'w') as file:
+    asset_url = response["external_url"]
+
+    with open(asset_file_path, "w") as file:
         file.write(f"[InternetShortcut]\nURL={asset_url}\n")
